+++ src/Presentations/Kin.KinHub.Shared.Api/GlobalUsings.cs
@@ -1,6 +1,9 @@
 global using Kin.KinHub.Shared.Api.AuthenticationFeature;
 global using Kin.KinHub.Shared.Api.Common;
+global using Kin.KinHub.Shared.Api.Common.Configuration;
+global using Kin.KinHub.Shared.Api.Common.Mcp;
 global using Kin.KinHub.Shared.Api.FamilyFeature;
+global using Kin.KinHub.Shared.Api.McpFeature.Contracts.Transport;
 global using Kin.KinHub.Shared.Api.RecipeFeature;
 global using Kin.KinHub.Shared.Api.RecipeAssistantFeature;
 global using Kin.KinHub.Core.Business.FamilyFeature;